+++ docs/theme/playground-config-2c5c2b01.js
@@ -1,0 +1,5 @@
+var colorNotebook = {
+    "playgroundWheels": ['micropip', 'pygments-2.19.2-py3-none-any.whl', 'coloraide-6.2.1-py3-none-any.whl'],
+    "notebookWheels": ['pyyaml', 'markdown-3.10-py3-none-any.whl', 'pymdown_extensions-10.20-py3-none-any.whl', 'micropip', 'pygments-2.19.2-py3-none-any.whl', 'coloraide-6.2.1-py3-none-any.whl'],
+    "defaultPlayground": "import coloraide\ncoloraide.__version__\nColor('red')"
+}